delete 2nd CommentService, update CommentDto
