--- src/main/java/faang/school/postservice/dto/CommentDto.java
@@ -1,22 +1,0 @@
-package faang.school.postservice.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@RequiredArgsConstructor
-@Data
-public class CommentDto {
-    private Long id;
-    private Long postId;
-    @NotBlank(message = "Comment content cannot be empty")
-    @Size(max = 4096, message = "Comment content is too long (max 4096)")
-    private String content;
-    @NotNull(message = "Author must be specified")
-    private Long authorId;
-    private LocalDateTime createdAt; // Readonly
-}
