+++ src/test/java/faang/school/postservice/util/service/comments/CommentServiceTest.java
@@ -1,7 +1,7 @@
 package faang.school.postservice.util.service.comments;
 
 import faang.school.postservice.client.UserServiceClient;
-import faang.school.postservice.dto.CommentDto;
+import faang.school.postservice.dto.comment.CommentDto;
 import faang.school.postservice.dto.user.UserDto;
 import faang.school.postservice.mapper.CommentMapper;
 import faang.school.postservice.model.Comment;
